requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.5
lxml==4.5.0
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
requests2==2.16.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.21.1